ip-whitelist/docker/requirements.txt
kubernetes==12.0.1
PyYAML==5.4.1
